move bot reactions to a module
