+++ go.mod
@@ -1,0 +1,7 @@
+module github.com/medzernik/SlovakiaDiscordBotGo
+
+go 1.16
+
+require (
+	github.com/bwmarrin/discordgo v0.23.2
+)
