+++ main.go
@@ -5,10 +5,10 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
-	"strings"
 	"syscall"
 
 	"github.com/bwmarrin/discordgo"
+	"github.com/medzernik/SlovakiaDiscordBotGo/responder"
 )
 
 // Variables used for command line parameters
@@ -34,9 +34,7 @@ func main() {
 	// In this example, we only care about receiving message events.
 	dg.Identify.Intents = discordgo.IntentsGuildMessages + discordgo.IntentsGuildMessageReactions
 
-	// Register the messageCreate func as a callback for MessageCreate events.
-	dg.AddHandler(messageCreate)
-	dg.AddHandler(reactionAdded)
+	responder.RegisterPlugin(dg)
 
 	// Open a websocket connection to Discord and begin listening.
 	err = dg.Open()
@@ -53,32 +51,4 @@ func main() {
 
 	// Cleanly close down the Discord session.
 	dg.Close()
-}
-
-// This function will be called (due to AddHandler above) every time a new
-// message is created on any channel that the authenticated bot has access to.
-func messageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
-
-	// Ignore all messages created by the bot itself
-	// This isn't required in this specific example but it's a good practice.
-	if m.Author.ID == s.State.User.ID {
-		return
-	}
-	// If the message is "ping" reply with "Pong!"
-	if m.Content == "Zasielkovna" {
-		s.ChannelMessageSend(m.ChannelID, "OVER 200% <a:medzernikShake:814055147583438848>")
-	}
-
-	// If the message is "pong" reply with "Ping!"
-	if m.Content == "pong" {
-		s.ChannelMessageSend(m.ChannelID, "Ping!")
-	}
-}
-
-func reactionAdded(s *discordgo.Session, mr *discordgo.MessageReactionAdd) {
-	fmt.Println("reactionAdded: Message received, comparing..")
-	if strings.Contains(strings.ToLower(mr.Emoji.Name), "kekw") {
-		s.MessageReactionAdd(mr.ChannelID, mr.MessageID, mr.Emoji.APIName())
-		fmt.Println("Trying to react")
-	}
 }
